it was added some validation on product and string connection mongo db online

+++ luiguiDev.Market.Business/ProductBusiness/ProductBusiness.cs
@@ -1,4 +1,5 @@
 ﻿using luigiDev.Market.DataAccess.ProductRepository;
+using luigiDev.Market.DataAccess.StoreRepository;
 using luigiDev.Market.Entities;
 using System;
 using System.Collections.Generic;
@@ -10,16 +11,24 @@ namespace luiguiDev.Market.Business.ProductBusiness
     public class ProductBusiness : IProductBusiness
     {
         readonly IProductService productService;
-        public ProductBusiness(IProductService productService)
+        readonly IStoreService storeService;
+        public ProductBusiness(IProductService productService, IStoreService storeService)
         {
             this.productService = productService;
+            this.storeService = storeService;
         }
 
-        public async Task<bool> CreateProductAsync(Product product)
+        public async Task<(bool created, string errorMessage)> CreateProductAsync(Product product)
         {
             try
             {
-                return await productService.CreateProductAsync(product);
+                if (await storeService.ExistsStore(product.StoreId))
+                {
+                    var storeCreated = await productService.CreateProductAsync(product);
+                    return (storeCreated, string.Empty);
+                }
+
+                return (false, "The store does not exist, please verify the store");
             }
             catch (Exception)
             {
@@ -28,11 +37,19 @@ namespace luiguiDev.Market.Business.ProductBusiness
             }
         }
 
-        public async Task<bool> DeleteProductAsync(Guid productoId)
+        public async Task<(bool deleted, string errorMessage)> DeleteProductAsync(Guid productoId)
         {
             try
             {
-                return await productService.DeleteProductAsync(productoId);
+                var Product = await productService.GetProductoByIdAsync(productoId);
+
+                if (Product != null)
+                {
+                    var deleted = await productService.DeleteProductAsync(productoId);
+                    return (deleted, string.Empty);
+                }
+
+                return (false, "The product that you want to delete, does not exist");
             }
             catch (Exception)
             {
@@ -80,11 +97,19 @@ namespace luiguiDev.Market.Business.ProductBusiness
             }
         }
 
-        public async Task<bool> UpdateProductAsync(Product product)
+        public async Task<(bool updated, string errorMessage)> UpdateProductAsync(Product product)
         {
             try
             {
-                return await productService.UpdateProductAsync(product);
+                var Product = await productService.GetProductoByIdAsync(product.ProductId);
+
+                if (Product != null)
+                {
+                    var updated = await productService.UpdateProductAsync(product);
+                    return (updated, string.Empty);
+                }
+
+                return (false, "The product that you want to update, does not exist");
             }
             catch (Exception)
             {
